requirements.txt
tensorflow-gpu==1.4.0
captcha==0.2.4
matplotlib==2.1.0